--- src/components/ListingsContainer/ListingsContainers.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import './ListingsContainer';
-
-const ListingsContainer = () => {
-  
-}
